refactor(pool-page): extract _tabs/oracle-tab
Move OracleTab (~340 lines) out of page.tsx into _tabs/oracle-tab.tsx.
Most complex tab: user-controlled sort (timestamp/oracleOk/oraclePrice/
priceDifference), search-aware fetch limits with desc-by-timestamp
override, dedicated chart query at 200 rows, and breach-banner.

Final tab extraction. page.tsx is now a slim shell:
  - Suspense + PoolDetail orchestrator
  - Re-exports for symbols imported by external tests
  - Conditional tab mounting (one-of-N pattern preserved — lazy-mount
    contract enforced by characterization tests in __tests__/page.test.tsx)

Behavior-preserving — verified via:
- pnpm test (1366/1366 green; oracle-tab tab search + sort tests pass)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=oracle vs baseline:
  zero structural changes, console clean (oracle data refreshes fast,
  so the diff is verbose data drift but no DOM-shape changes)

page.tsx: 894 → 559 lines (-335).

Total impact across 11 commits:
  page.tsx: 2831 → 559 lines (-2272, -80%).
